Extract out `biome_line_index` crate
