+++ crates/biome_lsp/Cargo.toml
@@ -21,6 +21,7 @@ biome_console        = { workspace = true }
 biome_deserialize    = { workspace = true }
 biome_diagnostics    = { workspace = true }
 biome_fs             = { workspace = true }
+biome_line_index     = { workspace = true }
 biome_lsp_converters = { workspace = true }
 biome_resolver       = { workspace = true }
 biome_rowan          = { workspace = true }
