+++ crates/biome_lsp/src/session.rs
@@ -9,7 +9,8 @@ use biome_console::markup;
 use biome_deserialize::Merge;
 use biome_diagnostics::PrintDescription;
 use biome_fs::BiomePath;
-use biome_lsp_converters::{PositionEncoding, WideEncoding, negotiated_encoding};
+use biome_line_index::WideEncoding;
+use biome_lsp_converters::{PositionEncoding, negotiated_encoding};
 use biome_service::Workspace;
 use biome_service::WorkspaceError;
 use biome_service::configuration::{LoadedConfiguration, load_configuration, load_editorconfig};
