+++ crates/biome_lsp_converters/src/from_proto.rs
@@ -1,7 +1,7 @@
-use crate::line_index::LineIndex;
-use crate::{LineCol, PositionEncoding, WideLineCol};
+use crate::PositionEncoding;
 use anyhow::{Context, Result};
-use biome_rowan::{TextRange, TextSize};
+use biome_line_index::{LineCol, LineIndex, WideLineCol};
+use biome_text_size::{TextRange, TextSize};
 use tower_lsp_server::lsp_types::{Position, Range};
 
 /// The function is used to convert a LSP position to TextSize.
